sadi/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-allauth==65.10.0
django-simple-history==3.10.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
#Python==3.12.10